# Improve performance of update-thread bound `RealmLive` retrievals