--- a/osu.Game/Database/RealmLive.cs
+++ b/osu.Game/Database/RealmLive.cs
@@ -129,7 +129,7 @@ namespace osu.Game.Database
         {
             Debug.Assert(ThreadSafety.IsUpdateThread);
 
-            if (dataIsFromUpdateThread)
+            if (dataIsFromUpdateThread && !data.Realm.IsClosed)
                 return;
 
             dataIsFromUpdateThread = true;