This is just an initial pass. I have some further potential improvements coming which build on this.

Also adds a structure for benchmarking and one minor safety addition I found while using `ILive` in more places.

Before:

|                                     Method | ReadsPerFetch |         Mean |       Error |      StdDev |     Gen 0 |     Gen 1 |  Gen 2 | Allocated |
|------------------------------------------- |-------------- |-------------:|------------:|------------:|----------:|----------:|-------:|----------:|
|                BenchmarkDirectPropertyRead |          1000 |   4,206.6 us |   106.01 us |   275.53 us |   85.9375 |   39.0625 |      - |  1,009 KB |
|    BenchmarkDirectPropertyReadUpdateThread |          1000 |   3,600.7 us |    70.84 us |   112.36 us |   89.8438 |   46.8750 | 7.8125 |    972 KB |
|             BenchmarkRealmLivePropertyRead |          1000 | 162,479.2 us | 3,106.41 us | 5,757.93 us | 3333.3333 | 1666.6667 |      - | 38,461 KB |
| BenchmarkRealmLivePropertyReadUpdateThread |          1000 | 162,876.8 us | 2,278.12 us | 2,130.95 us | 3250.0000 | 1500.0000 |      - | 38,456 KB |
|              BenchmarkDetachedPropertyRead |          1000 |   1,228.4 us |    24.08 us |    29.57 us |    7.8125 |    3.9063 |      - |     91 KB |

After:

|                                     Method | ReadsPerFetch |         Mean |       Error |      StdDev |     Gen 0 |     Gen 1 |  Gen 2 | Allocated |
|------------------------------------------- |-------------- |-------------:|------------:|------------:|----------:|----------:|-------:|----------:|
|                BenchmarkDirectPropertyRead |          1000 |   4,246.0 us |    83.35 us |   122.17 us |   85.9375 |   39.0625 | 7.8125 |  1,009 KB |
|    BenchmarkDirectPropertyReadUpdateThread |          1000 |   3,975.4 us |    78.83 us |    73.74 us |   85.9375 |   39.0625 |      - |    972 KB |
|             BenchmarkRealmLivePropertyRead |          1000 | 165,873.6 us | 3,202.16 us | 4,163.71 us | 3250.0000 | 1500.0000 |      - | 38,461 KB |
| BenchmarkRealmLivePropertyReadUpdateThread |          1000 |   3,903.6 us |    71.84 us |    67.20 us |   85.9375 |   39.0625 | 7.8125 |    972 KB |
|              BenchmarkDetachedPropertyRead |          1000 |   1,035.4 us |    20.63 us |    31.51 us |    3.9063 |    1.9531 |      - |     46 KB |