@peppy 2022-01-19
They indeed are. I saw these and then promptly forgot to fix them while dealing with markdown tables.