@bdach 2022-01-20
I have no idea why it is possible, but in the remaining test failures it looks like something is falling over quite badly. For instance, sometimes `s` inside the following read is somehow `null`:

https://github.com/ppy/osu/blob/1a207251628d50ae701ec73a4c47f0f27ca29930/osu.Game.Tests/Skins/TestSceneBeatmapSkinResources.cs#L30-L34

I've confirmed that in this case `RealmLive` is going through the new update-thread path, but no clue what is going on otherwise. Doesn't help that I can only reproduce that like 20% of the time without debugger, and pretty much never with debugger stuff enabled. The only reason I found out the above is via adding logging statements.