@peppy 2022-01-21
Was also seeing this locally. Will need some further investigation, I also didn't manage to get a debugger at the correct point - it always passes when attached.