@peppy 2022-01-21
@bdach see 81b5717ae793e334fdcf8efd72d20debfb49e9ca for the answer to the issue above. i won't fix on this PR because it will likely be obsoleted by that one. should immediately make sense as to why it was such a weird failure and why debug "fixed" it.